New functionality for omuxsock
a. Abstract unix socket names (including network namespaces)
b. Connected socket support (SOCK_STREAM, SOCK_SEQPACKET)
c. Support for load, and action configuration
d. Bug fix related to closing/reopening sockets

+++ tests/uxsockrcvr.c
@@ -31,6 +31,7 @@
 #include "config.h"
 #include <stdio.h>
 #include <stdlib.h>
+#include <stddef.h>
 #include <string.h>
 #include <unistd.h>
 #include <signal.h>
@@ -53,15 +54,25 @@
 #define DFLT_TIMEOUT 60
 
 char *sockName = NULL;
-int sock;
+int sockType = SOCK_DGRAM;
+int sockConnected = 0;
 int addNL = 0;
+int abstract;
+#define MAX_FDS 4
+int sockBacklog = MAX_FDS - 1;
+struct pollfd fds[MAX_FDS];
+int nfds;
 
 
 /* called to clean up on exit
  */
 void cleanup(void) {
-    unlink(sockName);
-    close(sock);
+    int index;
+
+    if (!abstract) unlink(sockName);
+    for (index = 0; index < nfds; ++index) {
+        close(fds[index].fd);
+    }
 }
 
 
@@ -75,10 +86,46 @@ void usage(void) {
             "usage: uxsockrcvr -s /socket/name -o /output/file -l\n"
             "-l adds newline after each message received\n"
             "-s MUST be specified\n"
+            "-a Use abstract socket name\n"
+            "-T {DGRAM|STREAM"
+#ifdef SOCK_SEQPACKET
+            "|SEQPACKET"
+#endif /* def SOCK_SEQPACKET */
+            "} Set socket type (default DGRAM)\n"
             "if -o ist not specified, stdout is used\n");
     exit(1);
 }
 
+static struct {
+    const char *id;
+    int val;
+    int connected;
+} _sockType_map[] = {
+    {"DGRAM", SOCK_DGRAM, 0},
+    {"STREAM", SOCK_STREAM, 1},
+#ifdef SOCK_SEQPACKET
+    {"SEQPACKET", SOCK_SEQPACKET, 1},
+#endif /* def SOCK_SEQPACKET */
+    {NULL, 0, 0},
+};
+
+static void _decode_sockType(const char *s) {
+    int index;
+
+    for (index = 0; _sockType_map[index].id; ++index) {
+        if (!strcasecmp(s, _sockType_map[index].id)) {
+            sockType = _sockType_map[index].val;
+            sockConnected = _sockType_map[index].connected;
+            return;
+        }
+    }
+    fprintf(stderr, "?Illegal socket type '%s'. Valid socket types:\n", s);
+
+    for (index = 0; _sockType_map[index].id; ++index) {
+        fprintf(stderr, "  %s\n", _sockType_map[index].id);
+    }
+    exit(1);
+}
 
 int main(int argc, char *argv[]) {
     int opt;
@@ -89,15 +136,18 @@ int main(int argc, char *argv[]) {
     struct sockaddr_un addr; /* address of server */
     struct sockaddr from;
     socklen_t fromlen;
-    struct pollfd fds[1];
+    int index;
 
     if (argc < 2) {
         fprintf(stderr, "error: too few arguments!\n");
         usage();
     }
 
-    while ((opt = getopt(argc, argv, "s:o:lt:")) != EOF) {
+    while ((opt = getopt(argc, argv, "as:o:lt:T:")) != EOF) {
         switch ((char)opt) {
+            case 'a':
+                abstract = 1;
+                break;
             case 'l':
                 addNL = 1;
                 break;
@@ -112,6 +162,9 @@ int main(int argc, char *argv[]) {
                 break;
             case 't':
                 timeout = atoi(optarg);
+                break;
+            case 'T':
+                _decode_sockType(optarg);
                 break;
             default:
                 usage();
@@ -135,7 +188,7 @@ int main(int argc, char *argv[]) {
     }
 
     /*      Create a UNIX datagram socket for server        */
-    if ((sock = socket(AF_UNIX, SOCK_DGRAM, 0)) < 0) {
+    if ((fds[0].fd = socket(AF_UNIX, sockType, 0)) < 0) {
         perror("server: socket");
         exit(1);
     }
@@ -145,39 +198,77 @@ int main(int argc, char *argv[]) {
     /*      Set up address structure for server socket      */
     memset(&addr, 0, sizeof(addr));
     addr.sun_family = AF_UNIX;
-    strcpy(addr.sun_path, sockName);
+    strncpy(addr.sun_path + abstract, sockName, sizeof(addr.sun_path) - abstract);
+    /*
+     * Abstract socket addresses do not require termination.
+     */
+    if (!abstract && addr.sun_path[sizeof(addr.sun_path) - 1]) {
+        addr.sun_path[sizeof(addr.sun_path) - 1] = 0;
+        fprintf(stderr, "warning: socket path truncated: %s\n", addr.sun_path);
+    }
 
-    if (bind(sock, (struct sockaddr *)&addr, sizeof(addr)) < 0) {
-        close(sock);
+    /*
+     * For pathname addresses, the +1 is the terminating \0 character.
+     * For abstract addresses, the +1 is the leading \0 character.
+     */
+    if (bind(fds[0].fd, (struct sockaddr *)&addr, offsetof(struct sockaddr_un, sun_path) + strlen(sockName) + 1) < 0) {
         perror("server: bind");
+        close(fds[0].fd);
         exit(1);
     }
 
-    fds[0].fd = sock;
+    if (sockConnected && listen(fds[0].fd, sockBacklog) == -1) {
+        perror("server: listen");
+        close(fds[0].fd);
+        exit(1);
+    }
+
     fds[0].events = POLLIN;
+    nfds = 1;
 
     /* we now run in an endless loop. We do not check who sends us
      * data. This should be no problem for our testbench use.
      */
 
     while (1) {
-        fromlen = sizeof(from);
-        rlen = poll(fds, 1, timeout);
+        rlen = poll(fds, nfds, timeout);
         if (rlen == -1) {
             perror("uxsockrcvr : poll\n");
             exit(1);
         } else if (rlen == 0) {
             fprintf(stderr, "Socket timed out - nothing to receive\n");
             exit(1);
-        } else {
-            rlen = recvfrom(sock, data, 2000, 0, &from, &fromlen);
-            if (rlen == -1) {
-                perror("uxsockrcvr : recv\n");
-                exit(1);
-            } else {
-                fwrite(data, 1, rlen, fp);
-                if (addNL) fputc('\n', fp);
+        }
+        if (sockConnected && fds[0].revents & POLLIN) {
+            fds[nfds].fd = accept(fds[0].fd, NULL, NULL);
+            fds[nfds].events = POLLIN;
+            fds[nfds].revents = 0;
+            if (fds[nfds].fd < 0) {
+                perror("accept");
+            } else if (++nfds == MAX_FDS) {
+                fds[0].events = 0;
             }
+        }
+        for (index = sockConnected; index < nfds; ++index) {
+            if (fds[index].revents & POLLIN) {
+                fromlen = sizeof(from);
+                rlen = recvfrom(fds[index].fd, data, 2000, 0, &from, &fromlen);
+                if (rlen == -1) {
+                    perror("uxsockrcvr : recv\n");
+                    exit(1);
+                } else {
+                    fwrite(data, 1, rlen, fp);
+                    if (addNL) fputc('\n', fp);
+                }
+            }
+            if (fds[index].revents & POLLHUP) {
+                close(fds[index].fd);
+                fds[index].fd = -1;
+                fds[index].events = 0;
+            }
+        }
+        while (nfds && fds[nfds - 1].fd < 0) {
+            --nfds;
         }
     }
